Edit: 프로젝트 구조 통일화
- 컴포넌트, 컨테이너 및 페이지로 분리
- 컴포넌트와 컨테이너의 세분화 된 폴더 별 index.js 로 일괄 Export
