+++ src/Router.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { Nav } from './components/Nav/Nav.component';
-import { SensorList } from './pages/SensorList/SensorList.page';
-import { GraphDashboard } from './pages/GraphDashboard/GraphDashboard.page';
+import { Nav } from './containers/index.containers';
+import { SensorList } from './pages/SensorList.page';
+import { GraphDashboard } from './pages/GraphDashboard.page';
 
 function Router() {
   return (
