+++ src/components/Graph/Graph.component.js
@@ -1,8 +1,7 @@
 import React, { Fragment, useEffect, useRef, useState } from 'react';
-import styled from 'styled-components';
 import { Line } from 'react-chartjs-2';
 
-export const TemperatureGraph = ({ channelData, feedData, options }) => {
+export const Graph = ({ field, channelData, feedData, options }) => {
   const [graphData, setGraphData] = useState();
   const chartRef = useRef();
 
@@ -14,8 +13,9 @@ export const TemperatureGraph = ({ channelData, feedData, options }) => {
         }),
         datasets: [
           {
-            label: channelData.field1,
-            data: feedData.map(feed => feed.field1),
+            label: channelData[`field${field}`],
+            data: feedData.map(feed => feed[`field${field}`]),
+            fill: false,
           },
         ],
       });
@@ -24,7 +24,12 @@ export const TemperatureGraph = ({ channelData, feedData, options }) => {
   if (graphData) {
     return (
       <Fragment>
-        <Line ref={chartRef} data={graphData} options={options} />
+        <Line
+          key={channelData.id}
+          ref={chartRef}
+          data={graphData}
+          options={options}
+        />
         <button onClick={() => chartRef.current.zoom(1.1)}>확대지롱</button>
         <button onClick={() => chartRef.current.zoom(0.9)}>축소지롱</button>
         <button onClick={() => chartRef.current.resetZoom()}>초기화지롱</button>
