--- src/components/Graph/index.js
@@ -1,4 +1,0 @@
-export * from './Calendar.component';
-export * from './HumidityGraph.component';
-export * from './PressureGraph.component';
-export * from './TemperatureGraph.component';
