--- src/components/Nav/index.component.js
@@ -1,2 +1,0 @@
-export * from './Breadcrumbs.component';
-export * from './Sidebar.component';
