--- src/pages/SensorList/components/index.js
@@ -1,1 +1,0 @@
-export * from './SensorTable.component';
