--- src/pages/SensorList/index.js
@@ -1,1 +1,0 @@
-export * from './components/index';
